+++ src/store/init.js
@@ -1,0 +1,7 @@
+import { Store } from 'tauri-plugin-store-api';
+
+/**
+ * Use to access the base store instance and its methods
+ * @type {Store}
+ */
+export const store = new Store('.artisan-settings');
