+++ src/store/operations/index.js
@@ -1,0 +1,6 @@
+export { addIllustration } from './illustration-add';
+export { updateIllustrationURL } from './illustration-update';
+export { removeIllustration } from './illustration-remove';
+export { addProject } from './project-add';
+export { updateProject } from './project-update';
+export { removeProject } from './project-remove';
